DTrOCR-15 - Adding helper functions and updating PRETRAINING_DATASET.md

+++ tools/copy_latin_fonts.py
@@ -9,7 +9,8 @@ from pathlib import Path
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(description='Script for copying latin fonts from source folder')
     parser.add_argument('--input', type=str, help='Path to Google Fonts folder')
-    parser.add_argument('--output', type=str, help='Path to copy font files')
+    parser.add_argument('--font_output', type=str, help='Path to copy font files to')
+    parser.add_argument('--list_output', type=str, help='Path to write font list to')
     args = parser.parse_args()
 
     font_files = glob.glob(os.path.join(args.input, '**', '*.ttf'), recursive=True)
@@ -24,6 +25,12 @@ if __name__ == "__main__":
             if 'subsets: \"latin\"' not in meta:
                 continue
 
-            output_file = Path(args.output) / Path(font_file).name
-            output_file.parent.mkdir(exist_ok=True, parents=True)
-            shutil.copy(font_file, output_file)
+            output_file = Path(args.font_output) / Path(font_file).name
+            if not output_file.is_file():
+                output_file.parent.mkdir(exist_ok=True, parents=True)
+                shutil.copy(font_file, output_file)
+
+    print("Writing font list")
+    with open(os.path.join(args.list_output, 'font_list.txt'), 'w') as file:
+        font_files = glob.glob(os.path.join(args.font_output, '*.ttf'))
+        file.write('\n'.join([Path(font_file).name for font_file in font_files]))
